Fix sq issues
